Changed scanline_timeout to float type / tmp in ppu.c is now uint8_t Signed-off-by: Rodrigo Javier Tobar Carrizo <rtobar@csrg.inf.utfsm.cl>

+++ include/ppu.h
@@ -55,7 +55,7 @@ typedef struct _ppu {
 	uint8_t spr_addr;   /* Address to be written by 0x2004 CPU RAM */
 
 	uint8_t mirroring;     /* Type of mirroring */
-	int scanline_timeout;  /* CPU cycles to next scanline */
+	float scanline_timeout;  /* CPU cycles to next scanline */
 	unsigned int lines;    /* Current scanline */
 	unsigned int frames;   /* Frame counting */
